pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
packages = [
  {include = "lm_service"},
  {include = "test"}
]
requires = ["poetry-core>=1.0.0"]

[tool.poetry]
authors = ["Alexander Belikov <alexander@growgraph.dev>"]
description = "Convert text to graphs. Create candidates. Entity linking"
name = "lm_service"
version = "0.5.2"

[tool.poetry.dependencies]
Flask = "^2.3.2"
Flask-RESTful = "^0.3.9"
Flask-SocketIO = "^5.2.0"
Unidecode = "^1.3.4"
aenum = "^3.1.11"
coreferee = "^1.4.0"
curlify = "^2.2.1"
dataclass-wizard = "^0.22.1"
flask-compress = "^1.13"
flask-cors = "^4.0.1"
flask-limiter = "^3.7.0"
lemminflect = "^0.2.2"
networkx = "^2.6.3"
pandas = "^2.0.3"
pathos = "^0.3.0"
pybind11 = "^2.12.0"
pygraphviz = "^1.6"
pyinflect = "^0.5.1"
pylatexenc = "^2.10"
pytest = "7.4.0"
python = ">=3.10,<3.12"
pyyaml = "^6"
requests = "^2.31.0"
simple-websocket = "^0.5.2"
spacy = {extras = ["cuda114"], version = "3.5"}
suthing = {git = "ssh://git@github.com/growgraph/suthing.git", rev = "v0.2.8"}
torch = "^1.13.0"
torchtext = "^0.12.0"
transformers = "^4.42.2"
viztracer = "^0.15.6"
waitress = "^2.1.2"
wordfreq = "^3.0.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
pylint = "^3.2.3"
